+++ packages/editor-common/web/src/Icons/index.js
@@ -25,4 +25,4 @@ export { default as TrashIcon } from './TrashIcon';
 export { default as WidthIcon } from './WidthIcon';
 export { default as CloseIcon } from './CloseIcon';
 export { default as EditIcon } from './EditIcon.svg';
-// export { default as InfoIcon } from './InfoIcon';
+export { default as InfoIcon } from './InfoIcon.svg';